+++ ProjectA/Assets/Scripts/Player.cs
@@ -5,7 +5,7 @@ using UnityEngine;
 public class Player : MonoBehaviour
 {
     public float speed;
-    public int hp;
-    public int armor;
+    public float hp;
+    public float armor;
     public Weapon weapon;
 }
